Adding Session service
