+++ app/src/pages/authPages/RiskGroup/index.js
@@ -8,7 +8,7 @@ import {
 } from 'react-native';
 import Button from '../../../components/UI/button';
 import styles from './styles';
-import userService from '../../../services/User';
+import SessionService from '../../../services/Session';
 import { Icon } from 'react-native-elements';
 import colors from '../../../../assets/styles/colorVariables';
 import riskGroups from '../../../utils/riskGroupsObject';
@@ -46,7 +46,7 @@ export default function RiskGroup({ route, navigation }) {
 
         try {
             setLoading(true);
-            await userService.signUp(completeRegistragionData);
+            await SessionService.signUp(completeRegistragionData);
             Alert.alert(
                 'Sucesso',
                 'Seu cadastro foi realizado com sucesso!',
