+++ app/src/pages/profile/ListProfile/index.js
@@ -12,6 +12,7 @@ import Button from '../../../components/UI/button';
 import { UserContext } from '../../../store/contexts/userContext';
 import actions from '../../../store/actions';
 import UserService from '../../../services/User';
+import SessionService from '../../../services/Session';
 import { Icon } from 'react-native-elements';
 import * as ImagePicker from 'expo-image-picker';
 import ConfirmationModal from '../../../components/modals/confirmationModal';
@@ -28,7 +29,7 @@ export default function Profile({ navigation }) {
     const [photo, setPhoto] = useState('');
 
     async function logout() {
-        await UserService.logOut();
+        await SessionService.signOut();
         dispatch({ type: actions.user.removeUserInfo });
     }
